+++ test/mix/tasks/extension/phoenix/pow.extension.phoenix.mailer.gen.templates_test.exs
@@ -78,6 +78,14 @@ defmodule Mix.Tasks.Pow.Extension.Phoenix.Mailer.Gen.TemplatesTest do
     end)
   end
 
+  test "warns no mailer templates" do
+    File.cd!(@tmp_path, fn ->
+      Templates.run(~w(--extension PowPersistentSession))
+
+      assert_received {:mix_shell, :info, ["Warning: No mailer view or template files generated for PowPersistentSession as no mailer templates has been defined for it."]}
+    end)
+  end
+
   describe "with extensions in env config" do
     setup do
       Application.put_env(:pow, :pow, extensions: Enum.map(@expected_template_files, &elem(&1, 0)))