+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v1.0.5 (TBA)
 
 * Added `extension_messages/1` to extension controllers and callbacks
+* Improved feedback for when no templates are generated for an extension with `mix pow.extension.phoenix.gen.templates` and `mix pow.extension.phoenix.mailer.gen.templates` tasks
 
 ## v1.0.4 (2019-03-13)